Refactor out ProductIdMappingConfiguration
